refactor(bench): extract the posture/assurance seam into its own module
The witness-arm work pushed the adapter package root and its test file past
their file-size ceilings. The gate's rule is that separable growth becomes a
module rather than more length on an oversized file, so:

* `stella_harbor/posture.py` — the engine posture, the assurance declaration,
  and the proof-stream fold. All pure functions of the model and the arm, with
  no Harbor/Docker/credential dependency, which is why the launcher, the
  manifest writer and the preregistration generator can all derive their hashes
  from one implementation instead of from copies that can disagree.
* `tests/test_posture.py` — the arm's tests, all three layers.

`posture.py` also had to be declared in `_FIXED_ADAPTER_SOURCE_PATHS`: the
claim launcher byte-compares an enumerated set of adapter sources against the
published commit, so an undeclared module would execute without ever being
compared. It refused the run until listed, which is the check working.

Ceilings for the irreducible residue updated via `make file-size-update` —
`__init__.py` +51 (run() wiring, metadata assembly, one method), and the two
secure-launcher files for the ambient-witness-author guard. `test_adapter.py`
tightened from 1735 to 1707.
